@@ -1,68 +1,71 @@
 {
   "name": "glee",
-  "version": "0.0.0",
+  "version": "2.0.0",
   "description": "Hapi boilerplate with glue, jest, ES6+ (stage-3), build scripts and everything you need to quick-start backend development with Flow type checking.",
   "engineStrict": true,
   "engines": {
-    "node": ">=6.9"
+    "node": ">=8.11.1"
   },
   "devDependencies": {
-    "aws-sdk": "^2.96.0",
+    "aws-sdk": "^2.229.1",
     "babel-cli": "^6.24.1",
-    "babel-core": "~6.24.1",
+    "babel-core": "~6.26.2",
     "babel-eslint": "^8.2.2",
     "babel-helpers": "^6.24.1",
-    "babel-jest": "~19.0.0",
+    "babel-jest": "~22.4.3",
     "babel-plugin-transform-async-to-generator": "^6.24.1",
-    "babel-plugin-transform-es2015-modules-commonjs": "^6.24.1",
+    "babel-plugin-transform-es2015-modules-commonjs": "^6.26.2",
     "babel-plugin-transform-flow-strip-types": "^6.22.0",
     "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-stage-3": "~6.24.1",
     "babel-register": "^6.24.1",
     "cfn": "^1.2.0",
     "chai": "^4.1.1",
-    "cucumber": "^2.3.1",
+    "cucumber": "^4.2.1",
     "eslint": "^4.10.0",
-    "eslint-config-airbnb-base": "^11.3.1",
+    "eslint-config-airbnb-base": "^12.1.0",
     "eslint-import-resolver-node": "^0.3.1",
     "eslint-plugin-flowtype": "^2.35.0",
     "eslint-plugin-import": "^2.7.0",
-    "flow-bin": "^0.50.0",
+    "flow-bin": "^0.70.0",
     "fs": "0.0.1-security",
     "gulp": "^3.9.1",
-    "gulp-clean": "^0.3.2",
+    "gulp-clean": "^0.4.0",
     "gulp-util": "^3.0.8",
     "gulp-zip": "^4.0.0",
     "husky": "^0.14.3",
-    "jest": "~19.0.2",
-    "lint-staged": "^4.0.3",
+    "jest": "~22.4.3",
+    "lint-staged": "^7.0.5",
     "pre-commit": "^1.2.2",
-    "prettier-eslint": "^6.4.2",
+    "prettier-eslint": "^8.8.1",
     "request": "^2.81.0"
   },
   "dependencies": {
     "awilix": "^3.0.7",
     "babel-runtime": "^6.25.0",
-    "blipp": "^2.3.0",
+    "blipp": "^3.0.0",
     "global": "^4.3.2",
-    "glue": "^4.2.0",
-    "good": "^7.2.0",
-    "good-console": "^6.4.0",
+    "glue": "^5.0.0",
+    "good": "^8.1.1",
+    "good-console": "^7.1.0",
     "good-file": "^6.0.1",
     "good-squeeze": "^5.0.2",
     "good-winston": "^4.0.0",
-    "hapi-auth-jwt2": "^7.3.0",
-    "hapi-swagger": "^7.7.0",
-    "inert": "^4.2.1",
-    "joi": "^10.6.0",
+    "hapi-auth-jwt2": "^8.1.0",
+    "hapi-dev-errors": "^2.1.0",
+    "hapi-swagger": "^9.1.1",
+    "inert": "^5.1.0",
+    "joi": "^13.2.0",
     "mototaxi": "^1.1.68",
     "pg": "^7.1.0",
     "pg-hstore": "^2.3.2",
     "sequelize": "^4.4.2",
-    "vision": "^4.1.1",
+    "vision": "^5.3.2",
     "winston": "^2.3.1",
-    "winston-loggly": "^1.3.1"
+    "winston-loggly": "^1.3.1",
+    "youch": "^2.0.7"
   },
   "scripts": {
     "clean": "rm -rf coverage dist tmp",
@@ -84,7 +87,9 @@
     "aat": "./node_modules/cucumber/bin/cucumber.js",
     "lint:staged": "lint-staged",
     "first-time": "cp ./src/config/sequelize.js.sample ./src/config/sequelize.js && cp ./src/config/loggly.js.sample ./src/config/loggly.js",
-    "docker:start": "docker-compose up --build",
+    "docker:start": "docker-compose up",
+    "docker:test": "docker-compose run web npm test",
+    "docker:test:watch": "docker-compose run web npm run test:watch",
     "docker:stop": "docker-compose down"
   },
   "author": "Byron Sommardahl <byron@acklenavenue.com> (http://acklenavenue.com)",
@@ -110,7 +115,7 @@
   },
   "lint-staged": {
     "*.js": [
-      "lint",
+      "eslint --fix",
       "git add"
     ]
   },
