@@ -1,23 +1,25 @@
 /* eslint-disable import/no-dynamic-require, global-require */
+import server from './server';
 
 const fs = require('fs');
 
-const getDirectoriesInPath = (p) =>
-  fs.readdirSync(p).filter((f) => fs.statSync(`${p}/${f}`).isDirectory());
-const getFilesInDirectory = (p) =>
-  fs.readdirSync(p).filter((f) => !fs.statSync(`${p}/${f}`).isDirectory());
+const getDirectoriesInPath = (p) => fs.readdirSync(p).filter((f) => fs.statSync(`${p}/${f}`).isDirectory());
+const getFilesInDirectory = (p) => fs.readdirSync(p).filter((f) => !fs.statSync(`${p}/${f}`).isDirectory());
 const getArrayFromFiles = (path) =>
   getFilesInDirectory(path)
     // $FlowFixMe
     .map((filename) => require(`${path}/${filename}`).default)
+    .map((plugin) => [].concat(plugin))
     .reduce((prev, curr) => prev.concat(curr));
 
-const connections = getArrayFromFiles(`${__dirname}/connections`);
+const infrastructurePlugins = getArrayFromFiles(`${__dirname}/plugins`);
 
-const infrastructurePlugins = getArrayFromFiles(`${__dirname}/registrations`);
-const routePlugins = getDirectoriesInPath(`${__dirname}/../routes`).map((d) => ({
-  plugin: `./routes/${d}`,
-}));
-const registrations = [].concat(infrastructurePlugins, routePlugins);
-
-export { connections, registrations };
+const routePlugins = getDirectoriesInPath(`${__dirname}/../routes`)
+  .map((d) => ({
+    route: `${__dirname}/../routes/${d}`,
+  }))
+  // $FlowFixMe
+  .map(({ route }) => require(`${route}`).default);
+const plugins: Array<any> = [].concat(infrastructurePlugins, routePlugins);
+const register = { plugins };
+export { server, register };
