@@ -1,4 +1,4 @@
-/* eslint-disable no-console, no-param-reassign */
+/* eslint-disable no-console, no-param-reassign, no-unused-vars */
 
 import * as mototaxi from 'mototaxi';
 import { AwilixResolutionError } from 'awilix';
@@ -6,35 +6,36 @@ import container from '../container';
 
 import { commandHandlers } from '../../domain';
 
-export function register(server: Object, options: Object, next: () => mixed) {
-  const logger = {
-    log: (message) => {
-      console.log(`mototaxi: ${message}`);
-    },
-  };
+exports.default = {
+  register: async (server: Object, options: Object) => {
+    const logger = {
+      log: (message) => {
+        console.log(`mototaxi: ${message}`);
+      },
+    };
 
-  const resolve = (handlerType) => {
-    try {
-      const resolved = container.resolve(`${handlerType.name}Handler`);
-      return resolved;
-    } catch (err) {
-      if (err instanceof AwilixResolutionError) {
-        console.log(err);
+    const resolve = (handlerType) => {
+      try {
+        const resolved = container.resolve(`${handlerType.name}Handler`);
+        return resolved;
+      } catch (err) {
+        if (err instanceof AwilixResolutionError) {
+          console.log(err);
+        }
+        throw err;
       }
-      throw err;
-    }
-  };
+    };
 
-  const config = {
-    logger,
-    commandHandlers,
-    resolve,
-  };
+    const config = {
+      logger,
+      commandHandlers,
+      resolve,
+    };
 
-  const dispatcher = mototaxi.getDispatcher(config);
-  server.app.dispatcher = dispatcher;
-
-  next();
-}
-
-exports.register.attributes = { name: 'mototaxi', version: '1.0.0' };
+    const dispatcher = mototaxi.getDispatcher(config);
+    server.app.dispatcher = dispatcher;
+  },
+  name: 'mototaxi',
+  version: '1.0.0',
+  options: {},
+};
