@@ -1,18 +1,16 @@
+import server from '../server';
+
 const Pack = require('../../package.json');
 
-export default [{
-  plugin: {
-    register: 'inert',
+export default [
+  {
+    plugin: 'inert',
   },
-},
-{
-  plugin: {
-    register: 'vision',
+  {
+    plugin: 'vision',
   },
-},
-{
-  plugin: {
-    register: 'hapi-swagger',
+  {
+    plugin: 'hapi-swagger',
     options: {
       info: {
         title: `${Pack.name} API Documentation`,
@@ -25,12 +23,10 @@ export default [{
         },
       },
       documentationPath: '/docs',
+      host: `localhost:${server.port}`,
     },
   },
-},
-{
-  plugin: {
-    register: 'blipp',
+  {
+    plugin: 'blipp',
   },
-},
 ];
